Declare `ignore` as default export
This fixes compatibility with TypeScript used in combination with ESM.
Currently, the only way to import `ignore` in TypeScript is through the
`import * as X from Y` construct, even though `ignore` is actually a
default export through the use of `module.exports = X`. When used in
combination with ESM, this would lead to `ignore` being imported as an
object with a `default` property.